requirements.txt
fastapi
uvicorn
twilio
python-dotenv
openai
python-multipart